+++ src/D3D11Renderer/Rendering/Resources/Buffers/D3D11UpdateBuffer.cs
@@ -1,4 +1,5 @@
-﻿using SharpDX.Direct3D11;
+﻿using SharpDX;
+using SharpDX.Direct3D11;
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
@@ -12,38 +13,127 @@ namespace Sovereign.D3D11Renderer.Rendering.Resources.Buffers
     /// Manages an updateable D3D11 resource buffer.
     /// </summary>
     /// 
-    /// Under the hood, this class actually manages two buffers - one
-    /// with D3D11_USAGE_DEFAULT that is used by the GPU, and one with
-    /// D3D11_USAGE_STAGING that is used to stage updates prior to use.
-    public sealed class D3D11UpdateBuffer<T> where T : unmanaged
+    /// Under the hood, this class manages a dynamic buffer. A local managed buffer
+    /// can be updated at any time; this buffer will be copied to the dynamic buffer
+    /// when the Update() method is called.
+    public sealed class D3D11UpdateBuffer<T> : System.IDisposable
+        where T : unmanaged
     {
 
-        public D3D11UpdateBuffer(D3D11Device device, int len)
-        {
+        /// <summary>
+        /// D3D11 device.
+        /// </summary>
+        private readonly D3D11Device device;
 
+        /// <summary>
+        /// GPU-side buffer.
+        /// </summary>
+        private Buffer gpuBuffer;
+
+        /// <summary>
+        /// Number of T objects in the buffer.
+        /// </summary>
+        public int Length { get; private set; }
+
+        /// <summary>
+        /// Local buffer for update.
+        /// </summary>
+        public T[] Buffer { get; private set; }
+
+        /// <summary>
+        /// GC handle for the local buffer.
+        /// </summary>
+        private GCHandle bufferHandle;
+
+        /// <summary>
+        /// Pointer to the beginning of the local buffer.
+        /// </summary>
+        private System.IntPtr bufferPtr;
+
+        /// <summary>
+        /// Length of the local buffer in bytes.
+        /// </summary>
+        private int bufferLenBytes;
+
+        /// <summary>
+        /// Creates a new updateable buffer.
+        /// </summary>
+        /// <param name="device">D3D11 device.</param>
+        /// <param name="bindFlags">D3D11 bind flags.</param>
+        /// <param name="sizeInT">Number of T objects in the buffer.</param>
+        public D3D11UpdateBuffer(D3D11Device device, BindFlags bindFlags, int sizeInT)
+        {
+            this.device = device;
+            Length = sizeInT;
+            CreateBuffers(bindFlags);
+            AllocateLocalMemory();
         }
 
-        private Buffer CreateSingleBuffer(D3D11Device device, BindFlags bindFlags, 
-            CpuAccessFlags cpuAccessFlags, ResourceUsage resourceUsage, 
-            int sizeInT, out GCHandle bufferHandle)
+        /// <summary>
+        /// Updates the buffer by copying the staging buffer to the GPU buffer.
+        /// </summary>
+        public void Update()
+        {
+            var context = device.Device.ImmediateContext;
+
+            /* Start by writing the local buffer to the buffer. */
+            context.MapSubresource(gpuBuffer, MapMode.WriteDiscard, MapFlags.None, out var stream);
+            stream.Write(bufferPtr, 0, bufferLenBytes);
+
+            /* Commit the staging buffer. */
+            context.UnmapSubresource(gpuBuffer, 0);
+        }
+
+        public void Dispose()
+        {
+            bufferHandle.Free();
+        }
+
+        /// <summary>
+        /// Creates the D3D11 buffers.
+        /// </summary>
+        /// <param name="bindFlags">D3D11 bind flags.</param>
+        private void CreateBuffers(BindFlags bindFlags)
+        {
+            gpuBuffer = CreateSingleBuffer(
+                bindFlags,
+                CpuAccessFlags.Write,
+                ResourceUsage.Dynamic);
+        }
+
+        /// <summary>
+        /// Allocates the local update buffer.
+        /// </summary>
+        private void AllocateLocalMemory()
+        {
+            Buffer = new T[Length];
+            bufferHandle = GCHandle.Alloc(Buffer, GCHandleType.Pinned);
+            bufferPtr = bufferHandle.AddrOfPinnedObject();
+            bufferLenBytes = Length * Marshal.SizeOf<T>();
+        }
+
+        /// <summary>
+        /// Creates a single D3D11 buffer.
+        /// </summary>
+        /// <param name="bindFlags">D3D11 binding flags.</param>
+        /// <param name="cpuAccessFlags">D3D11 CPU access flags.</param>
+        /// <param name="resourceUsage">D3D11 resource usage.</param>
+        /// <returns>D3D11 buffer.</returns>
+        private Buffer CreateSingleBuffer(BindFlags bindFlags, 
+            CpuAccessFlags cpuAccessFlags, ResourceUsage resourceUsage)
         {
             var desc = new BufferDescription()
             {
                 BindFlags = bindFlags,
                 CpuAccessFlags = cpuAccessFlags,
                 OptionFlags = ResourceOptionFlags.None,
-                SizeInBytes = sizeInT * Marshal.SizeOf<T>(),
+                SizeInBytes = Length * Marshal.SizeOf<T>(),
                 StructureByteStride = 0,
                 Usage = resourceUsage
             };
 
-            /* Pin the buffer storage so we can take a pointer. */
-            var data = new T[sizeInT];
-            bufferHandle = GCHandle.Alloc(data, GCHandleType.Pinned);
-            var dataPtr = bufferHandle.AddrOfPinnedObject();
-
             /* Create buffer. */
-            return new Buffer(device.Device, dataPtr, desc);
+            return new Buffer(device.Device, desc);
         }
 
     }